+++ DIMS-Core/DIMS-Core.BusinessLayer/Extensions/MiddlewareServiceExtensions.cs
@@ -19,7 +19,7 @@ namespace DIMS_Core.BusinessLayer.Extensions
             services.AddTransient<IUserService, UserService>();
             services.AddTransient<IMemberService, MemberService>();
             services.AddTransient<ITaskService, TaskService>();
-            services.AddTransient<ITaskTracksService, TaskTracksService>();
+            services.AddTransient<ITaskTrackService, TaskTrackService>();
 
             services.AddDatabaseDependencies()
                 .AddIndentityDependencies()
