+++ DIMS-Core/DIMS-Core/Views/TaskTrack/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<TaskTracksViewModel>
+﻿@model IEnumerable<TaskTrackViewModel>
 @{
     ViewData["Title"] = "My task traks";
 }
