+++ DIMS-Core/DIMS-Core/Views/_ViewImports.cshtml
@@ -3,6 +3,6 @@
 @using DIMS_Core.Models.Sample
 @using DIMS_Core.Models.Member
 @using DIMS_Core.Models.Task
-@using DIMS_Core.Models.TaskTracks
+@using DIMS_Core.Models.TaskTrack
 @using DIMS_Core.BusinessLayer.Models.Account
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
